[riscv] Fix Zbb simulator and disassembler error
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5619020
Auto-Submit: Yahan Lu <yahan@iscas.ac.cn>
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Commit-Queue: Yahan Lu <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#94358}

+++ test/cctest/test-assembler-riscv32.cc
@@ -408,7 +408,7 @@ UTEST_R1_FORM_WITH_RES(zexth, int32_t, int32_t, 0xB080, 0xB080)
 UTEST_R2_FORM_WITH_RES(rol, uint32_t, 16, 2, 64)
 UTEST_R2_FORM_WITH_RES(ror, uint32_t, 16, 2, 4)
 UTEST_I_FORM_WITH_RES(rori, int32_t, 16, 2, 4)
-UTEST_R1_FORM_WITH_RES(orcb, int32_t, int32_t, 0x10010011, 0xFFFF00FF)
+UTEST_R1_FORM_WITH_RES(orcb, int32_t, int32_t, 0x10010011, int32_t(0xFFFF00FF))
 
 // -- Memory fences --
 // void fence(uint8_t pred, uint8_t succ);
